MolecularFingerprints

Documentation for MolecularFingerprints.

MolecularFingerprints.MHFP

MolecularFingerprints.fingerprint

MolecularFingerprints.mhfp_hash_from_molecular_shingling

MolecularFingerprints.mhfp_shingling_from_mol

MolecularFingerprints.smiles_from_atoms

MolecularFingerprints.smiles_from_circular_substructures

MolecularFingerprints.smiles_from_rings

MolecularFingerprints.tanimoto

MolecularFingerprints.MHFP — Type

MHFP

Class for MHFP fingerprint generator/featurizer. Contains settings and parameters for MHFP fingerprint generation.

MolecularFingerprints.fingerprint — Method

fingerprint(mol::SMILESMolGraph, calc::MHFP{N})

Calculates the MHFP fingerprint of the given molecule and returns it as a bit vector

MolecularFingerprints.mhfp_hash_from_molecular_shingling — Method

mhfp_hash_from_molecular_shingling(shingling::Vector{String}, Encoder::MHFP)

Calculate the MinHash values from a given Molecular shingling.

The given encoder contains parameters such as the length of the hashed vector or the maximum permitted hash number, but also the seed of the rng, and the random numbers that were generated using this seed, that are used in the hashing process.

The algorithm is described in more detail in the original authors paper.

MolecularFingerprints.mhfp_shingling_from_mol — Function

mhfp_shingling_from_mol(
    mol::MolGraph,
    radius::Int = 3,
    rings::Bool = true,
    min_radius::Int = 1)

Calculate the "molecular shingling" of a given molecule.

A molecular shingling is a vector of "SMILES"-strings, calculated from the ring structures and atom types of the molecule (optional), and the circular substructures around each heavy (=non-hydrogen) atom of the molecule.

Arguments

mol::MolGraph: the molecule for which to calculate the shingling.

radius::Int=3: the radius up to which the substructures around each atom should be considered

rings::Bool=true: if true (default), extract the smallest set of smallest rings (sssr) of the molecule, and include the SMILES-string of each contained ring in the molecular shingling

min_radius::Int=1: minimum radius of the substructures around each atom to be considered. Default is 1. If min_radius=0, include the SMILES-string of each heavy (i.e., non-hydrogen) atom of the molecule is included in the molecular shingling.

MolecularFingerprints.smiles_from_atoms — Method

smiles_from_atoms(mol::MolGraph)

Return vector containing SMILES strings of all atoms of the given molecule.

MolecularFingerprints.smiles_from_circular_substructures — Method

smiles_from_circular_substructures(
    mol::MolGraph,
    radius::Int,
    min_radius::Int)

Return vector of SMILES strings of circular substructures around all atoms of a molecule.

For each atom of the given molecule, extract the substructures of radii min_radius to radius, and generate their corresponding SMILES strings.

MolecularFingerprints.smiles_from_rings — Method

smiles_from_rings(mol::MolGraph)

Return vector containing SMILES strings of all rings in the SSSR of the given molecule.

SSSR stands for the smallest set of smallest rings of the molecule.

Note: This function uses the function sssr from MolecularGraph.jl, which returns a "true" smallest set of smallest rings of the given molecule. However, in the original implementation of the mhfp algorithm, the "symmetrisized sssr" is used, which in some cases is non-minimal, i.e., contains an additional ring. The rdkit function to get the symmetrisized sssr is not available in MolecularGraph.jl or in RDKitMinimalLib, which is why the standard sssr is used. In most cases, this will not have any effect, but for some molecules, such as cubane, it will.

MolecularFingerprints.tanimoto — Method

tanimoto(a::BitVector, b::BitVector) -> Float64

Calculate the Tanimoto similarity coefficient (Jaccard Index) between two fingerprints. Formula: c / (a + b - c) where c is intersection count.